Update report engine charts bar group
