--- MvvX.Plugins.Open-XML-SDK/MvvX.Plugins.Open-XML-SDK/Word/Models/Charts/ChartCategorie.cs
@@ -1,11 +1,0 @@
-﻿namespace MvvX.Plugins.OpenXMLSDK.Word.Models.Charts
-{
-    public class ChartCategorie
-    {
-
-        public string Name { get; set; }
-
-        public string Color { get; set; }
-
-    }
-}
