--- MvvX.Plugins.Open-XML-SDK/MvvX.Plugins.Open-XML-SDK/Word/Models/Charts/ChartSerie.cs
@@ -1,22 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace MvvX.Plugins.OpenXMLSDK.Word.Models.Charts
-{
-    public class ChartSerie
-    {
-        /// <summary>
-        /// Name of the serie
-        /// </summary>
-        public string Name { get; set; }
-
-        /// <summary>
-        /// Color
-        /// </summary>
-        public string Color { get; set; }
-
-        /// <summary>
-        /// Values
-        /// </summary>
-        public List<string> Values { get; set; }
-    }
-}
