+++ MvvX.Plugins.Open-XML-SDK/MvvX.Plugins.Open-XML-SDK/Word/ReportEngine/Models/Charts/BarChartType.cs
@@ -1,0 +1,8 @@
+﻿namespace MvvX.Plugins.OpenXMLSDK.Word.ReportEngine.Models.Charts
+{
+    public enum BarChartType
+    {
+        BarChart = 1,
+        BarGroupingChart = 2
+    }
+}
