+++ MvvX.Plugins.Open-XML-SDK/MvvX.Plugins.Open-XML-SDK/Word/ReportEngine/Models/Charts/BarModel.cs
@@ -2,8 +2,13 @@
 
 namespace MvvX.Plugins.OpenXMLSDK.Word.ReportEngine.Models.Charts
 {
-    public class BarChartModel
+    public class BarModel : BaseElement
     {
+        /// <summary>
+        /// Type of the barchart
+        /// </summary>
+        public BarChartType BarChartType { get; set; }
+
         /// <summary>
         /// Graph Title
         /// </summary>
@@ -15,9 +20,34 @@ namespace MvvX.Plugins.OpenXMLSDK.Word.ReportEngine.Models.Charts
         public bool ShowTitle { get; set; }
 
         /// <summary>
+        /// Indicate if we show major grid lines
+        /// </summary>
+        public bool ShowLegend { get; set; }
+
+        /// <summary>
+        /// Indicate if we delete axis for values
+        /// </summary>
+        public bool DeleteAxeValue { get; set; }
+
+        /// <summary>
+        /// Space between line categories
+        /// </summary>
+        public int? SpaceBetweenLineCategories { get; set; }
+
+        /// <summary>
+        /// Indicate if we delete axis for categories
+        /// </summary>
+        public bool DeleteAxeCategory { get; set; }
+
+        /// <summary>
+        /// Legend font family
+        /// </summary>
+        public string FontFamilyLegend { get; set; }
+        
+        /// <summary>
         /// Categories
         /// </summary>
-        public IList<string> Categories { get; set; }
+        public List<BarCategory> Categories { get; set; }
 
         /// <summary>
         /// Taille par défaut des textes dans le graphique
@@ -26,13 +56,49 @@ namespace MvvX.Plugins.OpenXMLSDK.Word.ReportEngine.Models.Charts
         public double? FontSize { get; set; } = null;
 
         /// <summary>
+        /// Indicate if labels data will be showed
+        /// </summary>
+        public bool ShowDataLabel { get; set; } = true;
+
+        /// <summary>
+        /// Define if the graph has a border
+        /// </summary>
+        public bool HasBorder { get; set; }
+
+        /// <summary>
+        /// Max width of the graph
+        /// </summary>
+        public double? MaxWidth { get; set; }
+
+        /// <summary>
+        /// Max height of the graph
+        /// </summary>
+        public double? MaxHeight { get; set; }
+
+        /// <summary>
+        /// Color of data labels
+        /// </summary>
+        public string DataLabelColor { get; set; }
+
+        /// <summary>
         /// Values
         /// </summary>
-        public IList<BarSerie> Values { get; set; }
+        public List<BarSerie> Series { get; set; }
 
         /// <summary>
         /// Show / Hide Borders
         /// </summary>
         public bool ShowBarBorder { get; set; }
+
+        /// <summary>
+        /// Indicate if we show major grid lines
+        /// </summary>
+        public bool ShowMajorGridlines { get; set; }
+
+        public BarModel()
+            : base(typeof(BarModel).Name)
+        {
+            BarChartType = BarChartType.BarChart;
+        }
     }
 }
